visual-novel/openvino/image_generator.py
import os
from flask import Flask, request, jsonify
from flask_cors import CORS
import base64
import io
import numpy as np
from PIL import Image
import logging

# This is a placeholder for OpenVINO integration
# In a real implementation, you would import OpenVINO modules and load models
# try:
#     import openvino as ov
# except ImportError:
#     logging.error("OpenVINO not installed. Please install it with 'pip install openvino'")

app = Flask(__name__)
CORS(app)

# Configuration
MODEL_PATH = os.environ.get('MODEL_PATH', './models')
LOG_LEVEL = os.environ.get('LOG_LEVEL', 'INFO')

# Setup logging
logging.basicConfig(level=getattr(logging, LOG_LEVEL),
                   format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

# Placeholder for model loading
# In a real implementation, you would load your models here
def load_models():
    logger.info("Loading OpenVINO models from %s", MODEL_PATH)
    # Placeholder for model loading code
    # Example:
    # core = ov.Core()
    # model = core.read_model(os.path.join(MODEL_PATH, "model.xml"))
    # compiled_model = core.compile_model(model, "CPU")
    # return compiled_model
    return None

# Placeholder for image generation
def generate_image_from_prompt(prompt, scene_type="background"):
    logger.info("Generating image for prompt: %s (scene_type: %s)", prompt, scene_type)
    
    # In a real implementation, you would use your OpenVINO model to generate an image
    # For now, we'll create a placeholder image
    
    # Create a simple colored image based on scene_type
    width, height = 1280, 720
    
    if scene_type == "background":
        # Create a gradient background
        image = np.zeros((height, width, 3), dtype=np.uint8)
        for y in range(height):
            for x in range(width):
                # Create a blue-green gradient
                b = int(255 * (1 - y / height))
                g = int(255 * (x / width))
                r = int(100 * (y / height))
                image[y, x] = [r, g, b]
    elif scene_type == "character":
        # Create a simple character silhouette
        image = np.ones((height, width, 3), dtype=np.uint8) * 240  # Light gray background
        # Draw a simple character silhouette
        center_x, center_y = width // 2, height // 2
        # Head
        for y in range(center_y - 100, center_y - 20):
            for x in range(center_x - 40, center_x + 40):
                if (x - center_x) ** 2 + (y - (center_y - 60)) ** 2 < 40 ** 2:
                    if 0 <= y < height and 0 <= x < width:
                        image[y, x] = [100, 100, 100]
        # Body
        for y in range(center_y - 20, center_y + 120):
            for x in range(center_x - 50, center_x + 50):
                if abs(x - center_x) < 50 * (1 - (y - (center_y - 20)) / 140):
                    if 0 <= y < height and 0 <= x < width:
                        image[y, x] = [100, 100, 100]
    else:
        # Default plain image
        image = np.ones((height, width, 3), dtype=np.uint8) * 200  # Light gray
    
    # Convert numpy array to PIL Image
    pil_image = Image.fromarray(image)
    
    # Save to bytes
    img_byte_arr = io.BytesIO()
    pil_image.save(img_byte_arr, format='PNG')
    img_byte_arr.seek(0)
    
    # Encode as base64
    img_base64 = base64.b64encode(img_byte_arr.read()).decode('utf-8')
    
    return {
        "image_base64": img_base64,
        "prompt": prompt,
        "scene_type": scene_type
    }

# Routes
@app.route('/health', methods=['GET'])
def health_check():
    return jsonify({'status': 'ok'})

@app.route('/generate-image', methods=['POST'])
def generate_image_api():
    data = request.json
    prompt = data.get('prompt')
    scene_type = data.get('scene_type', 'background')
    
    if not prompt:
        return jsonify({'error': 'Prompt is required'}), 400
    
    try:
        result = generate_image_from_prompt(prompt, scene_type)
        return jsonify(result)
    except Exception as e:
        logger.exception("Error generating image")
        return jsonify({'error': str(e)}), 500

if __name__ == '__main__':
    # Load models on startup
    # models = load_models()
    
    # Start the Flask app
    app.run(host='0.0.0.0', port=8081, debug=True)