visual-novel/openvino/requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
numpy==1.23.5
pillow==9.3.0
openvino==2023.1.0